requirements.txt
Django==5.0.1
djangorestframework==3.14.0
python-dotenv==1.0.0
ollama>=0.1.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
email-validator==2.1.0
Pillow==10.2.0
langdetect==1.0.9
chardet==5.0.0

